Fall back on `sys.real_prefix` for Python 2.7
`sys.base_prefix` was not introduced until Python 3.3 which made this
hook raise an uncaught exception in Python 2.7. Catch this and fall back
to `sys.real_prefix` if it is set (i.e. we are in a virtual environment
-- this is set by `activate_this.py` when activating a virtual
environment from the `virtualenv` package) or `sys.prefix` otherwise.

Since this regards Python 2.7 compatibility, this should only be
relevant for the cx_Freeze 5.x branch.