Include doc directory in sdists
This includes the license file, which the license itself requires be included, as well as useful things for downstream packagers and users like the changelog.  Also, lots of downstream packagers build the documentation when it is available.

The directory appears to only include text files, so it shouldn't increase the size of the sdists too much.

Fixes #376.