Modify setup.py for building new base
The Win32GUI_GPU base is now being built by the setup build command.
A new linker flag is passed when building with the `msvc` compiler, and that
will cause the linker to emit the import library in the build temp directory.
This is beacuse we don't need to comsume the import library, and it does not
seem to be any way to stop the compiler emitting the library, unless `msbuild`
is used (which has a `/NOIMPLIB` command line argument.)