Add hook to importlib_metadata load.
importlib_metadata is a library which provides an API for accessing an
installed package’s metadata (see PEP 566), such as its entry points or
its top-level name. It works by finding *.egg or *.dist-info files from
installed packages. Because cx_freeze discards *.egg and *.dist-info
files on freezing, it breaks importlib_metadata.

Even worse, on import the package tries to extract its own metadata and
fails. Which makes it unloadable.

As a workaround, include .dist-info file of the package itself, allowing
it to be loaded. It won't function properly anyway because it can't find
any other package.