Only copy a file if should copy a file (#995)
Files marked in the exclusion list were being copied. Files were only excluded when they were detected as a dependency, however, when the entire folder was copied, unnecessary files were copied. Packages like pyqt5 and pyside2 have many duplicate files and were copied and generated a list of dependencies in another directory again.
Now, that is over, and I hope it does not cause regressions (at least it did not in my tests and my projects).